refactor(terminal): move the block jump plan and its bounds to Rust
`BlockJump` typed ghostty's i16 binding limit twice — once as the
re-anchor magnitude, once as the largest forward hop — and then chunked
`ordinal - 1` against it in a `while` loop at the call site. Chunking is a
decision: each positive `jump_to_prompt` re-counts from the new viewport
top's `down(1)`, so consecutive hops COMPOSE to the full delta and land
the exact prompt where a single saturated step lands short.

`slopdesk_terminal::blocks::jump_plan` answers the hops, and the two
bounds are stated once where the value lives. What is left in Swift is
the effect: issuing them. `maxStep` survives only as the door a test
asserts the bound through — nothing plans against it any more.

+++ Tests/SlopDeskWorkspaceCoreTests/Workspace/WB3BlockRoutingDispatchTests.swift
@@ -295,7 +295,7 @@ final class WB3BlockRoutingDispatchTests: XCTestCase {
     func testReAnchorDeltaFitsGhosttyI16BindingParameter() {
         XCTAssertLessThanOrEqual(
             BlockJump.reAnchorDelta,
-            Int(Int16.max) + 1,
+            UInt32(Int16.max) + 1,
             "-reAnchorDelta must parse as ghostty's i16 binding parameter",
         )
         XCTAssertGreaterThanOrEqual(
@@ -361,14 +361,14 @@ final class WB3BlockRoutingDispatchTests: XCTestCase {
     /// threshold is `> maxStep`, not `>=`.
     func testBlockJumpSingleHopAtBoundaryTwoHopsJustBeyond() {
         let single = RecordingSurfaceActions()
-        BlockJump.toPromptOrdinal(UInt32(BlockJump.maxStep + 1), using: single) // step = maxStep exactly
+        BlockJump.toPromptOrdinal(BlockJump.maxStep + 1, using: single) // step = maxStep exactly
         XCTAssertEqual(
             Array(single.actions[2...]), ["jump_to_prompt:\(BlockJump.maxStep)"],
             "a step of exactly maxStep is one in-range hop",
         )
 
         let split = RecordingSurfaceActions()
-        BlockJump.toPromptOrdinal(UInt32(BlockJump.maxStep + 2), using: split) // step = maxStep + 1
+        BlockJump.toPromptOrdinal(BlockJump.maxStep + 2, using: split) // step = maxStep + 1
         XCTAssertEqual(
             Array(split.actions[2...]),
             ["jump_to_prompt:\(BlockJump.maxStep)", "jump_to_prompt:1"],
